fix(war-room): drop zod from shared, inline schema in bot (pass-4)
Builder pass-3 confirmed pass-2's peerDependencies approach still
broke a fresh `web npm ci && npm run build`: web reaches the
shared package via a `file:../shared/war-room` symlink, and
Next.js's symlink-resolved module resolution walks the shared
package's real-path parents (none of which contain zod), not
web's hoisted node_modules. peerDependencies are a declaration
of intent, not a hoisting mechanism — npm doesn't install peer
deps into the consumer for file: deps.

Pass-4: keep the shared package zod-free entirely. Export only
the VERDICT_VALUES enum constant; each consumer (this commit:
bot synthesizer; PR 3c slice 2: web resolve-action route) builds
its own `z.enum(VERDICT_VALUES)` schema with its own zod install.
Two short z.object declarations vs. fragile package-resolution
surgery, and avoids the dual-zod runtime-instance hazard the
pass-2 smoke test was trying to mitigate.

Changes:

- shared/war-room/src/queen-verdict.ts: remove `import { z }`,
  remove `DerivedVerdictSchema` export, replace `z.infer<>` with
  a hand-written `DerivedVerdict` interface, document the
  zod-out-of-shared rationale in the module header.
- shared/war-room/package.json: drop the `peerDependencies: { zod }`
  declaration AND remove zod from devDependencies. The package
  is now zod-free at every layer (no peer, no dev, no runtime).
- shared/war-room/src/queen-verdict.test.ts: replace the
  DerivedVerdictSchema test block with a VERDICT_VALUES enum-pin
  test (the consumer's schema tests live with each consumer).
- shared/war-room/dist/*: regenerated.
- bot/api/lib/queen/verdict-deriver.ts: declare
  `DerivedVerdictSchema` locally using `z.enum(VERDICT_VALUES)` +
  bot's own zod install; existing tests import it from this file
  unchanged.
- web/package.json: drop `zod` from runtime deps (web has zero
  zod imports of its own; the resolve-action route in PR 3c
  slice 2 will add zod when it's introduced).
- web/src/server/queen-verdict-schema-smoke.test.ts: deleted.
  This test was the pass-2 dual-zod risk mitigation; that risk
  no longer exists since the shared package has no zod, so the
  test's preconditions are unsatisfiable.
- web/package-lock.json: regenerated to reflect zod removal.

Verified end-to-end:
  cd shared/war-room && rm -rf dist node_modules && npm ci && tsc -p tsconfig.json
  cd web            && rm -rf node_modules .next   && npm ci && npm run build
→ "Compiled successfully in 1484ms". Was "Module not found:
  Can't resolve 'zod'" pre-fix. shared/war-room: 255 tests pass.
web: 1530 tests pass. bot verdict-deriver: 37 tests pass.
