+++ shared/war-room/package.json
@@ -3,10 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "type": "module",
-  "description": "Shared war-room storage layer + redis-lock helper + queen-verdict primitives. Imported by web (HTTP route handlers) and bot (direct internal callers); not published to npm. The consumer's @upstash/redis and zod installs are reused by source-level resolution; this package declares them as peerDependencies (zod) and devDependencies (@upstash/redis) to avoid a dual-package hazard (two installations producing non-assignable instances).",
-  "peerDependencies": {
-    "zod": "^3.25.0 || ^4.0.0"
-  },
+  "description": "Shared war-room storage layer + redis-lock helper + queen-verdict primitives. Imported by web (HTTP route handlers) and bot (direct internal callers); not published to npm. The consumer's @upstash/redis install is reused by source-level resolution; this package declares it as a devDependency to avoid a dual-package hazard. zod is intentionally NOT a dep here — see queen-verdict.ts header for the file: + symlink resolution rationale.",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "exports": {
@@ -29,7 +26,6 @@
   "devDependencies": {
     "@types/node": "^22.0.0",
     "@upstash/redis": "1.37.0",
-    "typescript": "^5.9.0",
-    "zod": "^3.25.76"
+    "typescript": "^5.9.0"
   }
 }
