+++ web/package.json
@@ -19,8 +19,7 @@
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "react-markdown": "^10.1.0",
-    "yaml": "^2.8.2",
-    "zod": "^3.25.76"
+    "yaml": "^2.8.2"
   },
   "devDependencies": {
     "@types/node": "^22.0.0",
